refactor: [UF-68] 폴더의 이름을 수정하였습니다. #30
