+++ src/components/ui/index.ts
@@ -1,2 +1,2 @@
 export * from './Button';
-export * from './Sooner';
+export * from './Sonner';
